requirements.txt
aenum==3.1.15
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dataclasses-json==0.5.13
Deprecated==1.2.14
Flask==2.3.2
frozenlist==1.4.0
future==0.18.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.240
langsmith==0.0.14
line-bot-sdk==3.1.0
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pydantic==1.10.11
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.19
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.9.2
